move context commands over
