+++ src/command/alias.rs
@@ -4,8 +4,7 @@ use rustyline::completion::Pair as RustlinePair;
 
 use crate::{
     cmd::{exec_match, start_clap, Cmd},
-    completer,
-    config,
+    completer, config,
     env::Env,
     output::ClickWriter,
 };
